requirements.txt
streamlit==1.32.0
feedparser==6.0.10
beautifulsoup4==4.12.2
langchain-google-genai==0.0.11
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0